+++ prov/efa/Makefile.include
@@ -41,6 +41,8 @@ _efa_files = \
 	prov/efa/src/efa_domain.c \
 	prov/efa/src/efa_domain_util.c \
 	prov/efa/src/efa_fabric.c \
+	prov/efa/src/efa_fabric_util.c \
+	prov/efa/src/rdm/efa_rdm_fabric.c \
 	prov/efa/src/efa_mr.c \
 	prov/efa/src/efa_strerror.c \
 	prov/efa/src/efa_user_info.c \
@@ -99,6 +101,7 @@ _efa_headers = \
 	prov/efa/src/efa_device.h \
 	prov/efa/src/efa_domain.h \
 	prov/efa/src/efa_domain_util.h \
+	prov/efa/src/efa_fabric_util.h \
 	prov/efa/src/efa_errno.h \
 	prov/efa/src/efa_user_info.h \
 	prov/efa/src/efa_prov_info.h \
@@ -123,6 +126,7 @@ _efa_headers = \
 	prov/efa/src/efa_mmio.h \
 	prov/efa/src/rdm/efa_rdm_peer.h	\
 	prov/efa/src/rdm/efa_rdm_domain.h \
+	prov/efa/src/rdm/efa_rdm_fabric.h \
 	prov/efa/src/rdm/efa_rdm_cq.h \
 	prov/efa/src/rdm/efa_rdm_cntr.h \
 	prov/efa/src/rdm/efa_rdm_ep.h \
