+++ prov/efa/src/efa_fabric.c
@@ -20,26 +20,10 @@
 
 #include "efa_cq.h"
 #include "efa_domain.h"
+#include "efa_fabric_util.h"
 #include "rdm/efa_rdm_domain.h"
+#include "rdm/efa_rdm_fabric.h"
 #include "efa_prov_info.h"
-#ifdef EFA_PERF_ENABLED
-const char *efa_perf_counters_str[] = {
-	EFA_PERF_FOREACH(OFI_STR)
-};
-#endif
-
-
-static int efa_fabric_destruct_base(struct efa_fabric *efa_fabric)
-{
-	int ret;
-
-	ret = ofi_fabric_close(&efa_fabric->util_fabric);
-	if (ret)
-		EFA_WARN(FI_LOG_FABRIC,
-			 "Unable to close fabric: %s\n",
-			 fi_strerror(-ret));
-	return ret;
-}
 
 static int efa_fabric_close(fid_t fid)
 {
@@ -53,58 +37,6 @@ static int efa_fabric_close(fid_t fid)
 
 	free(efa_fabric);
 	return 0;
-}
-
-static int efa_rdm_fabric_close(fid_t fid)
-{
-	struct efa_rdm_fabric *rdm_fabric;
-	int ret;
-
-	rdm_fabric = container_of(fid, struct efa_rdm_fabric,
-				  efa_fabric.util_fabric.fabric_fid.fid);
-
-	ret = efa_fabric_destruct_base(&rdm_fabric->efa_fabric);
-	if (ret)
-		return ret;
-
-	if (rdm_fabric->shm_fabric) {
-		ret = fi_close(&rdm_fabric->shm_fabric->fid);
-		if (ret) {
-			EFA_WARN(FI_LOG_FABRIC,
-				"Unable to close fabric: %s\n",
-				fi_strerror(-ret));
-			return ret;
-		}
-	}
-
-#ifdef EFA_PERF_ENABLED
-	ofi_perfset_log(&rdm_fabric->perf_set, efa_perf_counters_str);
-	ofi_perfset_close(&rdm_fabric->perf_set);
-#endif
-	free(rdm_fabric);
-
-	return 0;
-}
-
-static int efa_non_cq_trywait(struct fid *fid)
-{
-	struct util_wait *wait;
-
-	switch (fid->fclass) {
-	case FI_CLASS_EQ:
-		wait = container_of(fid, struct util_eq, eq_fid.fid)->wait;
-		break;
-	case FI_CLASS_CNTR:
-		wait = container_of(fid, struct util_cntr, cntr_fid.fid)->wait;
-		break;
-	case FI_CLASS_WAIT:
-		wait = container_of(fid, struct util_wait, wait_fid.fid);
-		break;
-	default:
-		return -FI_EINVAL;
-	}
-
-	return wait->wait_try(wait);
 }
 
 static int efa_trywait(struct fid_fabric *fabric, struct fid **fids, int count)
@@ -125,35 +57,13 @@ static int efa_trywait(struct fid_fabric *fabric, struct fid **fids, int count)
 	return FI_SUCCESS;
 }
 
-static int efa_rdm_trywait(struct fid_fabric *fabric, struct fid **fids, int count)
-{
-	struct efa_cq *efa_cq;
-	int ret, i;
-
-	for (i = 0; i < count; i++) {
-		if (fids[i]->fclass == FI_CLASS_CQ) {
-			efa_cq = container_of(fids[i], struct efa_cq, util_cq.cq_fid.fid);
-			if (!efa_cq->util_cq.wait)
-				return -FI_EINVAL;
-			ret = efa_cq->util_cq.wait->wait_try(efa_cq->util_cq.wait);
-		} else {
-			ret = efa_non_cq_trywait(fids[i]);
-		}
-		if (ret)
-			return ret;
-	}
-	return FI_SUCCESS;
-}
-
 /*
- * Feature strings advertised by this build, per fabric.
+ * Feature strings advertised by the efa-direct fabric.
  *
  * Features are runtime-discoverable flags whose answer may differ
  * between efa-direct and efa (efa-rdm) because the two fabrics
- * exercise different code paths. efa_features_direct contains strings
- * valid for the efa-direct fabric; efa_rdm_features contains
- * strings valid for the efa fabric (RDM and DGRAM). A string absent
- * from the matching list returns false.
+ * exercise different code paths. A string absent from this list
+ * returns false.
  *
  * Consumers query by literal string, so renaming is a breaking change;
  * prefer adding new strings over editing existing ones.
@@ -162,53 +72,14 @@ static const char * const efa_features_direct[] = {
 	"mixed_hmem_iov",
 };
 
-static const char * const efa_rdm_features[] = {
-	/*
-	 * "mixed_hmem_iov" is not advertised on the efa (RDM) fabric:
-	 * efa_rdm_pke_copy_payload_to_ope() still dispatches copy
-	 * methods based on desc[0] only, so a multi-iov request mixing
-	 * host and HMEM descriptors can still misbehave even after the
-	 * send-path descriptor scans are in place.
-	 *
-	 * The NULL placeholder keeps this a valid initializer under
-	 * pre-C23 compilers (notably MSVC), which reject an empty {}.
-	 * efa_feature_in() skips NULL entries.
-	 */
-	NULL,
-};
-
-static bool efa_feature_in(const char * const *list, size_t n, const char *feature)
-{
-	size_t i;
-
-	if (!feature)
-		return false;
-
-	for (i = 0; i < n; i++)
-		if (list[i] && strcmp(list[i], feature) == 0)
-			return true;
-
-	return false;
-}
-
 static bool efa_fabric_feature_query_direct(const char *feature)
 {
 	return efa_feature_in(efa_features_direct,
 			      ARRAY_SIZE(efa_features_direct), feature);
 }
 
-static bool efa_rdm_fabric_feature_query(const char *feature)
-{
-	return efa_feature_in(efa_rdm_features,
-			      ARRAY_SIZE(efa_rdm_features), feature);
-}
-
 static struct fi_efa_feature_ops efa_feature_ops_direct = {
 	.query = efa_fabric_feature_query_direct
-};
-
-static struct fi_efa_feature_ops efa_rdm_feature_ops = {
-	.query = efa_rdm_fabric_feature_query
 };
 
 static int efa_fabric_ops_open(struct fid *fid, const char *ops_name,
@@ -216,18 +87,6 @@ static int efa_fabric_ops_open(struct fid *fid, const char *ops_name,
 {
 	if (strcmp(ops_name, FI_EFA_FEATURE_OPS) == 0) {
 		*ops = &efa_feature_ops_direct;
-		return FI_SUCCESS;
-	}
-
-	EFA_WARN(FI_LOG_FABRIC, "Unknown ops name: %s\n", ops_name);
-	return -FI_EINVAL;
-}
-
-static int efa_rdm_fabric_ops_open(struct fid *fid, const char *ops_name,
-				   uint64_t flags, void **ops, void *context)
-{
-	if (strcmp(ops_name, FI_EFA_FEATURE_OPS) == 0) {
-		*ops = &efa_rdm_feature_ops;
 		return FI_SUCCESS;
 	}
 
@@ -243,14 +102,6 @@ static struct fi_ops efa_fi_ops = {
 	.ops_open = efa_fabric_ops_open
 };
 
-static struct fi_ops efa_rdm_fi_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = efa_rdm_fabric_close,
-	.bind = fi_no_bind,
-	.control = fi_no_control,
-	.ops_open = efa_rdm_fabric_ops_open
-};
-
 static struct fi_ops_fabric efa_ops_fabric_direct = {
 	.size = sizeof(struct fi_ops_fabric),
 	.domain = efa_domain_open,
@@ -259,38 +110,6 @@ static struct fi_ops_fabric efa_ops_fabric_direct = {
 	.wait_open = ofi_wait_fd_open,
 	.trywait = efa_trywait
 };
-
-static struct fi_ops_fabric efa_rdm_ops_fabric = {
-	.size = sizeof(struct fi_ops_fabric),
-	.domain = efa_rdm_domain_open,
-	.passive_ep = fi_no_passive_ep,
-	.eq_open = ofi_eq_create,
-	.wait_open = ofi_wait_fd_open,
-	.trywait = efa_rdm_trywait
-};
-
-/*
- * Walk efa_util_prov.info, calling ofi_fabric_init for each entry until
- * something other than -FI_ENODATA is returned. The first matching info
- * configures @p efa_fabric->util_fabric. Shared by efa_fabric_open_base
- * and efa_rdm_fabric_open.
- */
-static int efa_fabric_init_base(struct efa_fabric *efa_fabric,
-				struct fi_fabric_attr *attr,
-				void *context)
-{
-	const struct fi_info *info;
-	int ret = -FI_ENODATA;
-
-	for (info = efa_util_prov.info; info; info = info->next) {
-		ret = ofi_fabric_init(&efa_prov, info->fabric_attr, attr,
-				      &efa_fabric->util_fabric, context);
-		if (ret != -FI_ENODATA)
-			break;
-	}
-
-	return ret;
-}
 
 static int efa_fabric_open_base(struct fi_fabric_attr *attr,
 				struct fid_fabric **fabric_fid, void *context)
@@ -316,43 +135,6 @@ static int efa_fabric_open_base(struct fi_fabric_attr *attr,
 
 err_free_fabric:
 	free(efa_fabric);
-	return ret;
-}
-
-static int efa_rdm_fabric_open(struct fi_fabric_attr *attr,
-			struct fid_fabric **fabric_fid, void *context)
-{
-	struct efa_rdm_fabric *rdm_fabric;
-	int ret;
-
-	rdm_fabric = calloc(1, sizeof(*rdm_fabric));
-	if (!rdm_fabric)
-		return -FI_ENOMEM;
-
-	ret = efa_fabric_init_base(&rdm_fabric->efa_fabric, attr, context);
-	if (ret)
-		goto err_free_fabric;
-
-#ifdef EFA_PERF_ENABLED
-	ret = ofi_perfset_create(&efa_prov, &rdm_fabric->perf_set,
-				 efa_perf_size, perf_domain, perf_cntr,
-				 perf_flags);
-	if (ret)
-		EFA_WARN(FI_LOG_FABRIC,
-			"Error initializing EFA perfset: %s\n",
-			fi_strerror(-ret));
-#endif
-
-	*fabric_fid = &rdm_fabric->efa_fabric.util_fabric.fabric_fid;
-	(*fabric_fid)->fid.fclass = FI_CLASS_FABRIC;
-	(*fabric_fid)->fid.ops = &efa_rdm_fi_ops;
-	(*fabric_fid)->ops = &efa_rdm_ops_fabric;
-	(*fabric_fid)->api_version = attr->api_version;
-
-	return 0;
-
-err_free_fabric:
-	free(rdm_fabric);
 	return ret;
 }
